refactor: Restructure source files and update import paths
- Created a new directory structure under `src/` for better organization: `src/core/`, `src/config/`, `src/ui/`, and `src/utils/`.
- Moved relevant files into their respective directories to enhance modularity.
- Updated import paths across the codebase to reflect the new structure.
- Added a new configuration loader in `src/config/config.ts` to manage application settings.
- Implemented AI content generation logic in `src/core/ai.ts` and Git operations in `src/core/git.ts`.
- Introduced prerequisite checks in `src/core/prerequisites.ts` to ensure required tools are installed.
- Enhanced user interface components in `src/ui/` for improved interaction during PR creation.
