+++ src/ui/cli.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env bun
 
 import { program } from "commander"; // Import commander
-import { version } from "../package.json"; // Import version for CLI
-import { registerGenerateCommands } from "./commands/generate";
-import { loadConfig } from "./config"; // Import config loader
+import { version } from "../../package.json"; // Import version for CLI
+import { registerGenerateCommands } from "../commands/generate";
+import { loadConfig } from "../config/config"; // Import config loader
 import { theme } from "./theme";
 
 /**
